Fix intermittent segfault when --logfile specified
Under the following conditions a segfault will
occur upon starting nwipe.

Run nwipe with --logfile option AND run nwipe in a directory that does
not have write permissions so the logfile can't be created AND run nwipe
without sudo or root or as as a user without sufficient privileges so
you don't have permissions to write to the directory.