Create_PDF_Certificate_Group
The code that handles nwipe.conf now checks
an existing nwipe.conf for the correct default
groups. This allows future additions to the
nwipe.conf file to be added to an existing nwipe.conf
file. This makes the code more robust.

The code that handles creation of groups, settings and
values has been made into a function so any future additions
to nwipe.conf can be added with a single call to this function.

i.e
nwipe_conf_populate( "Default_Wipe.Method", "PRNG" );

The code can now handle creation of groups to a arbitrary
depth of four levels, however currently we only use a
group depth of 1.

i.e
CCC :
{
 XXX :
 {
   YYY :
   {
     ZZZ :
     {
     };
     TTT :
     {
     };
   };
 };
};