Fixes the s shift s bug The s shift s bug is reported here https://github.com/PartialVolume/shredos.x86_64/issues/301
To summarize, if no drives are selected and then the user presses
s (lower case) a warning appears indicating that the user
should press S (upper case) to start the wipe. This
warning appears for about 3 seconds but during this time if the
user presses S (upper case) nwipe would immediately complete,
having wiped no drives and requesting the user to press the
spacebar to exit. The is incorrect behaviour.
The bug doesn't appear if the user pressed S after the 3
seconds elapsed and the warning message disappeared.

This patch fixes this so that it does not exit but displays the
warning for 3 seconds and then waits for input.