Reformat layout of PRNG help text.
The help text for all prngs was formatted
so that it is consistent in terms of layout
and positioning with the help text
as displayed by the methods help text.
Specifically so that it still remains
legible with a 80x25 legacy (nomodset)
terminal.

Also added the ability to run the prng
benchmark from within the prng selection
screen.